Bugfix: move importing bits from sspmod_janus_EntityController to sspmod_janus_Importer and make it raise the memory limit to 512M during parsing.
